Fix link to github repo for 'Edit this page' link
At the bottom of the Introduction page, https://realworld-docs.netlify.app/docs/intro, the link to the github repo seems to be broken.

It is sending us to:

https://github.com/gothinkster/realworld/tree/main/docs/docs/docs/intro.mdx

I think it should be:

https://github.com/gothinkster/realworld/blob/main/apps/documentation/docs/intro.mdx

So, I have made a pull request here. I am not sure if this is the correct place for this change to be made.

I notice that there may be a similar problem a few lines below with the link to the blog on line 33, but I could not find the blog so I have not investigated further.